Fix minor compile warning from Xcode.

+++ src/emucore/tia/frame-manager/FrameManager.cxx
@@ -139,7 +139,8 @@ void FrameManager::setState(FrameManager::State state)
       notifyFrameComplete();
 
       if (myTotalFrames > Metrics::initialGarbageFrames)
-        myJitterEmulation.frameComplete(myCurrentFrameFinalLines, myVsyncEnd - myVsyncStart);
+        myJitterEmulation.frameComplete(myCurrentFrameFinalLines,
+            static_cast<Int32>(myVsyncEnd - myVsyncStart));
 
       notifyFrameStart();
